+++ core-bundle/src/Image/Studio/Figure.php
@@ -60,7 +60,6 @@ class Figure
      * All arguments but the main image result can also be set via a Closure
      * that only returns the value on demand.
      *
-     * @param ImageResult                                                                 $image          Main image
      * @param Metadata|(\Closure(self):Metadata|null)|null                                $metadata       Metadata container
      * @param array<string, string|null>|(\Closure(self):array<string, string|null>)|null $linkAttributes Link attributes
      * @param LightboxResult|(\Closure(self):LightboxResult|null)|null                    $lightbox       Lightbox
@@ -317,9 +316,15 @@ class Figure
             $templateData['href'] = $href;
             $templateData['attributes'] = ''; // always define attributes key if href is set
 
-            // Map "imageTitle" to "linkTitle"
-            $templateData['linkTitle'] = ($templateData['imageTitle'] ?? null) ?? StringUtil::specialchars($metadata->getTitle());
-            unset($templateData['imageTitle']);
+            // Use link "title" attribute for "linkTitle" as it is already output explicitly in image.html5 (see #3385)
+            if (\array_key_exists('title', $linkAttributes)) {
+                $templateData['linkTitle'] = $linkAttributes['title'];
+                unset($linkAttributes['title']);
+            } else {
+                // Map "imageTitle" to "linkTitle"
+                $templateData['linkTitle'] = ($templateData['imageTitle'] ?? null) ?? StringUtil::specialchars($metadata->getTitle());
+                unset($templateData['imageTitle']);
+            }
         } elseif ($metadata->has(Metadata::VALUE_TITLE)) {
             $templateData['picture']['title'] = StringUtil::specialchars($metadata->getTitle());
         }