+++ core-bundle/src/Resources/contao/widgets/SelectMenu.php
@@ -37,6 +37,17 @@ class SelectMenu extends Widget
 	protected $strTemplate = 'be_widget';
 
 	/**
+	 * @param array $arrAttributes
+	 */
+	public function __construct($arrAttributes=null)
+	{
+		parent::__construct($arrAttributes);
+
+		$this->preserveTags = true;
+		$this->decodeEntities = true;
+	}
+
+	/**
 	 * Add specific attributes
 	 *
 	 * @param string $strKey